requirements.txt
anna-protocol-sdk==1.2.13
web3==6.11.3
python-dotenv==1.0.0